pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.23",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.3",
]
